add extensible command support for REPL
