+++ golem-rib-repl/src/command/mod.rs
@@ -1,12 +1,22 @@
+pub use registry::*;
+pub use untyped::*;
+
+mod builtin;
+mod registry;
+mod untyped;
+
 use crate::rib_context::ReplContext;
 
-/// A trait representing a REPL command that can:
-/// - Parse user input from a string
-/// - Execute logic based on that input
-/// - Print results or errors to the user
+/// A Command implementation will do the following:
+///  - Parse user input from a string that follows the command which is of the pattern `:command-name input` and gets a structured Input
+///  - The Input will soon require a Display for documentation purpose
+///  - The Input will be executed to get a structured Output or an ExecutionError
+///  - Print results or errors to the user
 ///
-/// Commands are invoked via a `:command-name` in a REPL, and everything following the
-/// command name is passed as raw input to `parse`.
+///
+///  To register new commands to the REPL from any client (Ex: golem-cli),
+///  create a struct that implements the `Command` trait and register it using `let mut registry = CommandRegistry::default()`,
+///  and `registry.register(MyCommand);`, and pass it the config when bootstrapping the REPL.
 pub trait Command {
     /// The structured input type resulting from parsing the raw REPL string.
     type Input;
@@ -26,12 +36,9 @@ pub trait Command {
     /// - `prompt_input`: The raw string entered by the user after the command name in the REPL.
     ///   For example, if the user types `:my-command foo bar`, then `prompt_input` will be
     ///   `"foo bar"`.
-    /// - `repl_context`: Shared context that may include state, configuration, or environment
-    ///   data needed during parsing.
-    ///
-    /// # Returns
-    /// - `Ok(Self::Input)` if parsing is successful.
-    /// - `Err(Self::InputParseError)` if the input is malformed or invalid.
+    /// - `repl_context`: An immutable projection of internal ReplState.
+    ///   This gives access to printer, current session of rib script etc
+    //
     fn parse(
         &self,
         prompt_input: &str,
@@ -42,11 +49,7 @@ pub trait Command {
     ///
     /// # Parameters
     /// - `input`: The structured input previously returned by `parse`.
-    /// - `repl_context`: The REPL context, providing necessary shared state for execution.
-    ///
-    /// # Returns
-    /// - `Ok(Self::Output)` if execution is successful.
-    /// - `Err(Self::ExecutionError)` if execution fails.
+    /// - `repl_context`: An immutable projection of internal ReplState
     fn execute(
         &self,
         input: Self::Input,
@@ -57,20 +60,20 @@ pub trait Command {
     ///
     /// # Parameters
     /// - `output`: The result returned by `execute` if it completed successfully.
-    /// - `repl_context`: The REPL context, providing necessary shared state for execution.
+    /// - `repl_context`: An immutable projection of internal ReplState
     fn print_output(&self, output: &Self::Output, repl_context: &ReplContext);
 
     /// Prints an error that occurred during input parsing.
     ///
     /// # Parameters
     /// - `error`: The error returned by `parse` when the user input is invalid.
-    /// - `repl_context`: The REPL context, providing necessary shared state for execution.
+    /// - `repl_context`: An immutable projection of internal ReplState
     fn print_input_parse_error(&self, error: &Self::InputParseError, repl_context: &ReplContext);
 
     /// Prints an error that occurred during command execution.
     ///
     /// # Parameters
     /// - `error`: The error returned by `execute` when something goes wrong during execution.
-    /// - `repl_context`: The REPL context, providing necessary shared state for execution.
+    /// - `repl_context`: An immutable projection of internal ReplState
     fn print_execution_error(&self, error: &Self::ExecutionError, repl_context: &ReplContext);
 }
