+++ golem-rib-repl/src/command/registry.rs
@@ -1,0 +1,38 @@
+
+use std::collections::HashMap;
+use std::sync::Arc;
+use crate::command::builtin::TypeInfo;
+use crate::UntypedCommand;
+
+#[derive(Default)]
+pub struct CommandRegistry {
+    commands: HashMap<String, Arc<dyn UntypedCommand>>,
+}
+
+impl CommandRegistry {
+    pub fn built_in() -> Self {
+        let mut registry = Self::default();
+        registry.register(TypeInfo);
+        registry
+    }
+
+    pub fn get_commands(&self) -> Vec<String> {
+        self.commands.keys().map(|name| name.to_string()).collect()
+    }
+
+    pub fn register<T>(&mut self, command: T)
+    where
+        T: UntypedCommand + 'static,
+    {
+        let name = command.name().to_string();
+        self.commands.insert(name, Arc::new(command));
+    }
+
+    pub fn get_command(&self, name: &str) -> Option<Arc<dyn UntypedCommand>> {
+        let result = self.commands.get(name);
+        match result {
+            Some(command) => Some(command.clone()),
+            None => None,
+        }
+    }
+}
